feat: add onExportToBackend prop so host can handle it
Move onExportToBackend to excalidraw app
